requirements.txt
click==8.0.4
python-editor==1.0.4
dearpygui==1.4.0
